inline RunnerAndConfigurationSettings.schemeManager (IDEA-233142 WIP)
this makes it a little bit more clear for API users that there are 3 types of run configuration sharing: StoreInArbitraryFileInProject, store in .idea/runConfigurations and store in workspace.xml. Perhaps, RunnerAndConfigurationSettings.isShared is still a bit confusing, so API needs to be improved further.
(this is work-in-progress: Ability to store run configurations in arbitrary `*.run.xml` files anywhere within content roots)

GitOrigin-RevId: 1f24b52e9e6f31144ba71bdd9d51ffcbca305771