[Gradle] Use build root project model explicitly to request 'build' level tooling models IDEA-228368
If the root project doesn't explicitly used when requesting models then the project which path was selected (for the project opening) will be the default.
It's possible to use(wrongly?) some non-root project for a Gradle multi-project opening. In that case it might cause issues in 'build' level model builders which expects build root project instead.

GitOrigin-RevId: fad5db67184f7bb8fde9325a359752f1c4b82c06