Remove Import Project action from the Welcome Screen, rename "Open" to "Open or Import": IDEA-222368
"Import" is not needed anymore, since "Open" is able to handle most of import features automatically. The action is going to stay in the main menu, at least, for a while, to cover cases which might be missing from the "Open" implementation.

"Open" is renamed to "Open or Import" to help people who used to the "Import Project" action.

GitOrigin-RevId: 288b89a9ca26d884c5a061db53291d8941b54b95