Constructor references with type parameters are not recognized as PsiMethodReferenceExpression
StatementParser used to ignore type parameters to constructor references
and it prevents the tokens of such statements from being grouped under
the PsiMethodReferenceExpression node in the PSI tree.

Since constructor references can have type parameters one might look
like a variable declaration. The parser didn't take this fact into
account and fails due to "::" after closing ">" of type arguments.

This patch adds a new check in StatementParser#parseStatement after
the attempt of parsing of a declaration statement fails to see if the
type declaration follows by "::" which is a marker that the examined
line might be a constructor reference.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>

GitOrigin-RevId: 968f4884c45c922ce2fd6a8e3614cca01423d5b3