deprecate projectComponentsInitialized in favour of white-listed EP
Another step to protect loading: `ProjectLifecycleListener.projectComponentsInitialized` is deprecated in favour of a new white-listed EP `ProjectServiceContainerInitializedListener`. As for `ProjectServiceContainerCustomizer` (that was introduced for legacy project model), new EP can be extended only by core plugin (runtime error for non-core).

Message bus must be not used for any control events, that’s why EP is used here (as for any such solutions in our new code). Even more — EP is a more simpler and suitable here. EP allows you to control order of execution — yes, if you want specific order, it means that your code is bad. But not possible for now to fix RunManager.

GitOrigin-RevId: 2493a14678d8604193dfff84008882d5ec4d2a24