Restore Java 8 programming language checks in Android Studio
There are some IntelliJ inspections around Java 8 programming
language recommendations. These would tell users to change
constructs to Java 8 language features, which would not work
on all API levels. Therefore, back in the day, we removed these
from the JavaAnalysisPlugin, and instead registered subclasses
of these in the Android plugin, where we *first* checked the
minSdkVersion to see if the inspection should be run.

Later, IntelliJ came up with a query mechanism for this, where
the inspections themselves can check whether each langauge
feature is available (and we implement that check in the
Android plugin).

However, when the subclasses were deleted, it looks like the
original inspections were not re-enabled (that, or they may
have been re-disabled at some point through a misunderstanding.)

Anyway, this adds the inspections back in.

Test: The inspections are already tested by IntelliJ unit tests
      and the feature lookup for Android has existing tests
      in the Android plugin.
Change-Id: If913be621aa913fa096e82245726049b9843b0ce