[ui designer] fix resolving of property references in GUI forms
ResourceBundleKeyReference::resolve() method always returned null if message bundle is located not in the root package. It seems that this bug didn't affect users because form files aren't show in text form, and "Find Usages" worked properly because it uses 'isReferenceTo' method. However fixing this is needed to implement an action which analyzes usages of property keys in modules (for IDEA-209382).

GitOrigin-RevId: 7292fea6410569393cb64dcab6c16231b2c20eeb