move CommonBundle from util to core (IDEA-203640)
Strings in CommonBundle.properties are used for buttons in dialogs so they should be used in IDE modules only. It would be even more natural to move the class to intellij.platform.ide module, but currently Task class from intellij.platform.core module uses some strings from CommonBundle so let's move it to core module for a start. This is enough to fix IDEA-203640.

GitOrigin-RevId: e5d14f4f0783d90e750730927019c8368fe27e2e