SSR: simplify matcher, only use internal match context and pattern
Prevents confusion between internal match context and passed in options and patterns, because options and patterns can no longer be passed in through methods. It is now more clear when a new matcher needs to be created: every pattern has its own matcher. This also means that SSBasedInspection no longer operates on match contexts, but uses and caches matchers. This fixes a case of duplicate highlights in the inspection.

GitOrigin-RevId: 6e52fef819d5e6bbb02f7cc9a56b1f56968d8c45