[vcs-log] use hash from the indexers in the index id
Theoretically, indexed roots could change with project roots staying the same, for example, when new indexers are implemented for other vcs-es. After this change, index files for old roots won't be deleted.

This commit is going to affect projects with both git and mercurial roots, since index id is going to change and old index files won't be deleted. However, this transition will most likely coincide with switching from one version to another, so it should not be a problem.

GitOrigin-RevId: 6131752a1338f2c16ae24e2df2d3e49152b8666e