[IDEA-232518] The balloon panel of a breakpoint is not FocusCycleRoot
The panel of Balloon used to define a FocusTraversalPolicy which was
supposed to define a custom focus traversal policy, but it did not do
the job very well: the TAB key was prevented from moving the focus to
elements of the Balloon panel.

This patch makes the balloon panel a FocusCycleRoot which makes it more
comfortable to use the keyboard and removes the custom focus traversal
policy since it is not needed anymore.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>