Auto-import settings on first start, don't show the dialog: IDEA-199232
The import settings dialog is not very pretty, is confusing, and is an obstacle on the way to the startup. At the same time, in most cases the default proposal is the best one. Therefore we remove the dialog in favor of intelligent import of the settings:

* If there are settings in well-known places, we choose the best match (the last modified one, roughly), and just import it right away.
* If there are settings in the cloud, import them.
* If there are no settings to import, but this is the first start of any JB IDE on this computer, we start with clean settings.
* And only in the case when this is not the first start, but there are no settings in well-known places, then we show the import dialog.

We decide that it is the very first start on this machine, if the user agreement has been just shown to the user.

A special VM option is added for QA to be able to test IDE on different settings.

GitOrigin-RevId: 39e30605b396e64928e09a46455ac2332d31e3cc