[light-git] do not create git status widget during tests
Widgets are instantiated by OCCodeInsightTestCaseBase.enableStatusBarWidgetsForTest for testing, however method isCompatibleWith is not called as the ide frame is null. Since it's not desirable to create LightGitStatusBarWidget out of light edit, as it triggers instantiation of services such as LightGitTracker, this commit adds an additional check whether the provider project is a light edit project.

GitOrigin-RevId: aa7fc972d72c16d4bf8f5aa21a3a5c2744f49aa9