Setup a JDK for a project which was automatically setup from sources: IDEA-222368, IDEA-233066
If there is any JDK available locally on disk, let's find it and use for the project. It will suit in most cases, and is much better, than showing a "Setup JDK" notification allowing the user to perform the same action manually.

At the same time, if no JDK is found locally, searching for it in the internet seems to be not desirable in this scenario: in this case the editor notification will be shown, as it was previously.

GitOrigin-RevId: 78975af5c52a3ce18c619e467cf80e3a5586d24e