Fix import old configs on Windows: IDEA-233219
On Windows, "historic" config location is ~/.IntelliJIdea2019.3/config.

Before 17eb4db1, we were checking if a directory is a config directory by its name (it should have started with the product prefix: "IntelliJIdea" in this case), but after that commit we began using a more precise isConfigDirectory method, which searches for 'options/other.xml' and other files.

However, since in Windows historic locations the configs were actually located not directly under the "config parent", but under a subfolder 'config' (there were '~/.IntelliJIdea2019.3/config/' and '~/.IntelliJIdea2019.3/system/'), we need to adjust isConfigDirectory to look into the 'config' subfolder as well.

GitOrigin-RevId: f2a06f7281206634bb3db31a820347f95aaf58f9