PY-40470 Move Python support from the bundled Grazie plugin to PyCharm core
First of all, it's easier to maintain the corresponding extensions this way as
the language evolves.

Also, it will help to extract remaining language-specific logic from Grazie to
other language plugins, as existing direction of optional dependencies (Grazie -> Python)
prevented doing the same for JavaScript because of an introduced plugin dependency
(JavaScript -> Grazie -> Python -> JavaScriptDebugger).

GitOrigin-RevId: a1e5a3d8121daf4f949abd0a602eb79c8a6f6dd9