[followup] IDEA-231063 [jcef] implement custom scheme handler for "file://" protocol
My original interpretation of the JCEF's suggestion that the "file:" problem should be solved by adding a "custom scheme handler" was incorrect. So I read it as a handler for "a custom scheme". Then I followed the javadoc for CefApp.onRegisterCustomSchemes(CefSchemeRegistrar registrar) that says: "Register a custom scheme. This method should not be called for the built-in HTTP, HTTPS, FILE, FTP, ABOUT and DATA schemes." And so I'd implemented a custom scheme replacing "file:" and a handler for it. Actually, the correct interpretation should be: a custom handler for the standard "file:" scheme. So I did it in this fix.

GitOrigin-RevId: d7fa2616bd53839f980565d98f8612783b115268