Use a Supplier to get EditorComposite's focusComponent
EditorComposite was setting the preferred focus component at construction
time by passing FileEditor#getPreferredFocusedComponent to the instance
of MyComponent it creates. Later on, when we request focus on the editor,
we will delegate to that component passed when creating MyComponent.

This would work fine if the preferred focus component never changes, but
that's not the case for some FileEditor implementations. For example, in
TextEditorWithPreview we return a different preferred focus component
depending on the current editor mode. In order to get the correct focus
component, we need to call FileEditor#getPreferredFocusedComponent again
whenever requesting focus on the EditorComposite.

GitOrigin-RevId: 9bdf1deb091e504aeaf0a0f2033be723b14b3616