[git-branches-dashboard] move field declaration before init block
Focus traversal policy is saved in the BRANCHES_UI_FOCUS_TRAVERSAL_POLICY field, and then used in the initMainUi method called from the init block. But since the field declaration was placed after the init block, the field was not initialized and the focus traversal policy for main component was not set. That led to focus in Vcs Log tab behaving in an unexpected way.

IDEA-233502

GitOrigin-RevId: 8e7ee382001876415a3b97532403217f470a1fa3