IDEA-CR-58797: Revert recent changes to the ConfigImportHelperTest
The changes were needed because of 17eb4db1, where we changed the algorythm which checks if a directory is a config directory. However, the changes appeared to be not correct (createConfigDir is used to create both existing and new config directories), and since 17eb4db1 has been reverted, the test can be reverted as well.

(cherry picked from commit 4353e42d6b9d3524f7dc7da0ee7e95383426d3c9)

GitOrigin-RevId: ee48bf92453d324044e526a0cd388770dd6892c2