[light-edit] only allow LightEditCompatible widget factories in Light Edit
Instead of making each widget factory to check for Light Edit in the isAvailable method, return back to the whitelist. Each widget should explicitly specify whether it is available in Light Edit, or it won't be shown (but all widgets are available by default in the main IDE).

This partially reverts 452bc319b87cd9feb8006273238ae765e1dff917.

GitOrigin-RevId: ec5ea3e2cb51a3b3e0af5a128b57ff05f9182871