[git-index] introduce vfs for index files
GitIndexFileSystem is introduced to allow viewing and editing index content directly. As doing so requires running git commands, some of the operations could not be performed synchronously since it requires running git in edt. To deal with this problem, read and write on disk are performed in background and are protected by a lock (separate for each file). Each file has cached information kept in an AtomicReference, containing hash, length and executable flag.

Refresh is done in two stages: CachedData for refreshed files is retrieved in background and then assigned in write action. Write is done in three stages: "before" events are sent in write action, actual write is performed in background outside of write action, "after" events are sent in write action.

GitOrigin-RevId: 7fe99dd3e52b3208a1d38707a877b5d84831f599