[intellij-sh] Simplify run before logic and get rid of ShBeforeRunTaskProvider
If we use `RunBefore` for scripts, we shouldn't execute them in the terminal. For achieving this
we introduce our own ShBeforeRunTaskProvider but it forces a user to add configuration only via
our provider and all scripts added via `Run Another Configuration` will fail.
EP runConfigurationBeforeRunProviderDelegate was found to help with achieving our goal. And thereby
we can rid of ShBeforeRunTaskProvider.

GitOrigin-RevId: 96aa968037cd5a549607c589651301afa31ad7d1