IDEA-234374 vcs: fix sorting in Local Changes tree
When comparing 'collapsed' FilePath nodes we should use fileName
of the _leading_ directory, and not of the _trailing_ one.

Ex: when sorting /src/test/zebra vs /test/alpaca we should
compare 'src' with 'test' and not 'zebra' with 'alpaca'.

Forcing ChangesBrowserNode comparator to take parent node into account
would be complicated (as we want different order for 'flat' and 'tree' modes).
Instead, collapse directories in a sorted tree while preserving nodes order.

GitOrigin-RevId: 56f5d72e20410745ee8afb04e8c02549ef64d396