[tabs] refactor TabbedContent
Introduce new concepts of TabGroupId and TabDescriptor to represent a tab group and individual tabs respectively. This allows to track relevant information like display name and disposable together instead of keeping them in the properties of the content or component.

Disposing process is simplified after this change: TabDescriptor is a Disposable itself, and it is registered on the content instance either in the ContentUtilEx or in the TabbedContentImpl. When the tab group is split or merged, TabDescriptor is re-registered on the new content before the old content gets disposed. When the tab is removed from TabbedContentImpl it gets disposed manually.

GitOrigin-RevId: 626e87ec027134be8b0311595262aa5bdc20cce7