IDEA-229503 Display javadoc rendered representation in editor
Fix incorrect behaviour on Enter pressed while caret is above the rendered doc (described in the ticket).

This makes caret 'not notice' the presence of fold region with empty placeholder. Previously caret could be located at
'both sides' of such a fold region, and moving from one side to another consumed arrow key press. Now there's only one
position of caret at such a fold region, and it maps to the region's start offset (we could choose to map it to the end
offset, but the former solution fits current implementation of rendered javadocs better).

GitOrigin-RevId: 79d0e4a75570dcf57306b298f8139c967f83a20b