Fix support for CREATE_EMPTY_DIRECTORIES
When a source folder already exists in a content entry but not in
the file system, a duplicate source root is created.

This does not seem to happen in the way IDEA uses
CREATE_EMPTY_DIRECTORIES. We find it very useful in Android Studio
tests, but it produces duplicate source folders if importData
runs for an existing but empty project.