Fix support for CREATE_EMPTY_DIRECTORIES
When a source folder already exists in a content entry but not in
the file system, a duplicate source root is created.

This does not seem to happen in the way IDEA uses
CREATE_EMPTY_DIRECTORIES. We find it very useful in Android Studio
tests, but it produces duplicate source folders if importData
runs for an existing but empty project.

close #1313

(cherry picked from commit fe188e5ac6053c4eb125cb70f9332215798b14f2)

Change-Id: I7c687b071c9c61d908c92b6e929dc424878d0015