IDEA-CR-59462 Don't mark dir for test project as "Source root" for RubyMine tests
Breaks tests:
 * `RailsCompletionTest#testAllHelpersCompletionTest`
 * Probably other tests...

In RubyMine we don't have such concept as "Source root". But in tests
we have. Because of that we have different configuration in testing
environment and in production environment. This could lead to tests
which are green but feature actually doesn't work (This is what
actually happened with
`RailsCompletionTest#testAllHelpersCompletionTest`).

GitOrigin-RevId: f0b144c7ef9c5c548cafd74ddbdba991d72f5df9