IDEA-CR-65062: PY-41000 Update PyPI cache on startup if the underlying file is missing
Previously we took into account only the timestamp of the last update, not
considering the file's actual presence. This timestamp could have been copied
from the previous IDE version configs leading to the cache being absent for
up to the whole update timeout (one day). Ideally, we should not keep such
data in settings and rely on the file's own modification time instead.

GitOrigin-RevId: 38ae8f53c31889ab1233a8c1204e2af6b2a7fd6f