Provide ability to start IDE on clean configs: IDEA-233305
Add an action "Restore Default Settings" to start with clean configs, teach the "Import Settings" action to be able to choose not only an exported "settings.zip" file, but also a config directory.

Both actions create a marker file "migrate.config" with a path with original config dir to migrate from (import settings) or without any content (restore default), and restart the IDE. On the next start the IDE reads the file and overwrites current config if needed.

A backup is created before overwriting settings. To support the case when settings are migrate from the backup dir, current settings are backed up first to a temp directory, and then moved to the standard backup directory after migration.

GitOrigin-RevId: cc2b9ba9f620162f3a95324eac4856739311a496