Moving lint from builder-model to new lint-model
This series of CLs migrates lint off of the builder-model API
and over to a new dedicated lint-model API which contains just
the parts of the builder-model that lint uses, and makes it
possible to evolve it independently. More importantly, it
supports serialization and we can therefore decouple building
the model in Gradle and having lint consume it. This unblocks
the instant execution work in Gradle.

Test: New and existing unit tests
Fixes: 151963156
Change-Id: I101470d46d097b9aaa5e7126f68877d11dd4ddaf