Don't suggest Conda environments in interpreter inspection quick fix
The inspection is run under a read action while the list of available
Conda environments is populated by an external process which we are
not allowed to run inside a read action.

We might consider switching from an inspection to a service that shows
a balloon notification about problems with a Python interpreter. But
for the moment the detection of Conda environments works only for:

* Opening new folders without .idea/
* Inside the Add Interpreter dialog

GitOrigin-RevId: 328cd1a096431a1b74806e2c70951212132ef337