Symbol: wrap PsiElements into Symbol even if they already implement Symbol, extract PsiElement from wrapped Symbols only
This simplifies expectations and helps to reduce the amount of client code.
Wrapped elements appear in parts where PsiElement-s were before, e.g. returning an element from PsiReference#resolve.
If the implementor migrates to the Symbol API, then the returned element should become a Symbol, and it should be
returned from the new #resolveReference method. This way we differ the elements from old APIs and symbols from the new
APIs without implying the relation between them.

Example: some client wasn't fully migrated, the element implements Symbol but returned from the old method.
In this case we wrap returned element and unwrap it to pass into old APIs.
If the client finishes migrating; and the element will be returned from new #resolveReference as a Symbol,
then we won't extract PsiElement form such Symbol, and won't pass it into old APIs.

GitOrigin-RevId: f190a302a222f8c98937d8159464d326b61040d1