[vcs-log] highlight non-picked commits in "Compare" tab
"Compare" Log ui uses a range filter instead of a branch filter and does not allow to set the filter manually. So when the branches chooser is displayed it can not filter the graph by the selected branch. Instead of allowing the user to select the branch, highlighter needs to use inclusive ref from the range.

(cherry picked from commit d45dcd37fac4d663b0e359934fd3ddcd40cae0b3)

GitOrigin-RevId: 097c50168fea7d86e5f6fe29bc297bb7ed5ffc54