[build scripts] remove confusing unused parallelStream call
'each' method called here is from Groovy library, and it converts Stream to iterator before calling it, so actually plugins are build sequentially. And it isn't possible to build them in parallel with the current implementation anyway, because AntBuilder used inside is not thread-safe, so it'll fail with exception like "module doesn't support the nested "layout" element" (IDEA-CR-58477).

GitOrigin-RevId: 9d36751051a7e16cf1b3011a96658424de1a68d8