[large files] IDEA-234404 IDEA-234864 Problematic interaction of PSI with Large Files Editor (another fix)
Untying of VirtualFile from Document (changes in LargeFileEditorImpl) will remove problematic interaction of PSI with Large Files Editor, but it will also lead to blocking of registration of caret moves in Navigation History in Document opened in Large File Editor. So, corresponding changes in EditorModel were made to recover this registration.

GitOrigin-RevId: 099b95eba70888a3b40d062bfc39b157f76d81ab