remove leading slash from default live templates
It works fine with a leading slash if templates are loaded by a plugin class loader. And this is what happens IRL. But it works differently in case of the Gradle Intellij Plugin. If plugins are declared as dependencies, they are loaded using the application class loader that can only find templates without the leading slash. We should definitely do something with with the Gradle plugin but for now let's avoid leading slashes.

GitOrigin-RevId: 3f63ef3708ab5f8d7d06143b0b2e97cff1af9510