cr-IDEA-1192: EA-220456 - do not search for SDKs from EDT (spoil - we almost do not)
ensure
SdkType#suggestHomePath does not do any heavy operations (it is used to show the file chooser dialog)

In fact, we use SdkType#suggestHomePaths for most of the cases, and it is called from a background thread.

This commit fixes the delegation - it is only allowed for the SdkType#suggestHomePaths to call the SdkType#suggestHomePath, not the opposite order. There were several cases of opposite delegation.

Also related to PY-41218

(cherry picked from commit d5716e81a35d33368812689eaadf4c11e2858e4f)

GitOrigin-RevId: 3a594e97c6c337c4ffafe19e8e6d90e4b24b3048