[workspace model] Store uncommitted modules in `LegacyBridgeModifiableModuleModel`
It's not the clear solution but worth considering that prior to this there was no
opportunity to get the uncommitted linked module from `OrderEntryViaTypedEntity`.
Module manager stores only committed modules and from now has a collection to store
uncommitted but managed outside of it. Why It controlled from the outside? Because
there is no clear way to get LegacyBridgeModifiableModuleModel but for module
manager we have such.

GitOrigin-RevId: b6a83077b297a4e7060c1b9d0c3ffc50460fbae6