Symbol [impl] don't wrap PsiElement#getReferences into TargetEvaluatorAwareReference
TargetEvaluatorAwareReference was supposed to adapt references returned from PsiElement#getReferences to resolve into
something which would be returned by TargetElementUtil later. This doesn't work since some evaluators actually depend on
the fact that PsiFile#findReferenceAt (which delegates to PsiElement#getReferences in the end) may return
multi-reference, meaning there are several references under the caret, and the evaluator chooses one.
In the new approach each reference is resolved separately, so when the reference from PsiElement#getReferences is
wrapped to go through TargetElementUtil it still doesn't know that there might be another reference at the same offset,
so the evaluator fails.

After the change TargetElementUtil is called when we are actually searching for targets instead of inside the wrapped
references.

GitOrigin-RevId: d344517cfe580aa37ff9378e81ca6ae1352a20f7