Don't use lazy methods to obtain tasks
Using these withType methods can cause ConcurrentModificationExceptions
as it is possible that the list returned to modified during the body of
the loops. This is since the returned collection is lazy and doesn't
force the creation of tasks. Instead we use a different overload of the
withType method that accepts a closure which ensures that all tasks of
the given type have already been created.

Currently it is possible for these loops to miss tasks of the wanted
type.

close #1332

(cherry picked from commit 2acd218bc6337d706628b8ec886482c55939649d)

Change-Id: I725ae052935707dd9502ef107d281bdc5a244772