[platform] raise minimum target Java version to 1.6 in IntelliJ project modules (IDEA-224536)
This is needed to simplify migration to Java 11, because it doesn't support cross-compilation to version 1.5 or older. This change means that users won't be able to run processes which use classes from IDE distribution (e.g. run JUnit tests, deploy to an application server, etc) under JDK 1.5 or older, they'll need to use JDK 1.6 for that.

GitOrigin-RevId: 1df45d51af993022778253fa13cd79d56090cd26