PY-40423: List ``LOCALAPPDATA`` to find python from store, do not use ``Get-AppxPackage`` cmdlet.
Appx apps are installed in special folder inside of ``LOCALAPPDATA``. You can get their location via cmdlet, but this path is not accessible for user in modern versions of Win10.

The official way is to use reparse point inside of ``LOCALAPPDATA``. Since it is added to ``PATH``, it should be accessible and executable.

However, this file can't be checked with ``File.exists`` and other ``File`` tools because it is a special reparse point.

* We list ``LOCALAPPDATA`` to find Python
* Appx-specific knowledge is in WinAppxTools.kt now

GitOrigin-RevId: 9eeaf186379c30531159b9fe4a2395930d4993c0