RUBY-16609 consume all comments at the start and end of a body/compound statement
This ensures all comments are in the same location in the PSI tree for a code block. The previous behaviour of having some leading/trailing comments outside of the compound statement meant a lot of hacks had to be implemented in order to determine what anchor comments are attached to, and also for moving/manipulating comments.

GitOrigin-RevId: 1aab3c87605245575e7ac462693782dcf5d2acb2