[codeinsight] IDEA-236413 Java: incorrect text for unused public fields with Inject
IDEA used did not use to check the visibility of a field and if there is
the javax.inject.Inject or similar annotations added to the field IDEA
used to always reported the the Private field is not used.

This patch adds a new method that generates the error messages
based on the visibility level of the field.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 89d7e2f783b1e54b239d15a2007d9e7376ffdbfa