PY-38249 Use SimpleColoredComponent to render entries in the package browser
Apart from cleaner implementation, it also gives us better control over what's
going to be copied in the clipboard - every SimpleColoredComponent has a notion
of "main" text JBList already relies on.

We can't use the standard ColoredListCellRenderer for this purpose because
the package name and its repository URL have different alignment in a row.

GitOrigin-RevId: bf6560c051d4acf0f80413b51c7ecb880419451c