use SrcOver compositing in editor by default
Previously Src compositing was used by default on Windows and macOS. It's hard to tell why it was enabled back in 2006 -
there are no known reasons to use it, and it makes it harder to draw transparent elements in editor (composite has to be
switched back and forth). There doesn't seem to be any performance impact from this change.

GitOrigin-RevId: 7ddb72a4445cf59f80c8cc6219fdf291bedb38be