[git-index] collect statuses only for dirty files
Instead of refreshing the whole repositories, statuses only for specific files are refreshed. Those files are collected into DirtyScope instances, one scope per repository root. Both git status and "unsaved" status are collected for dirty files. This prevents from getting duplicated statuses in the tree for the unsaved files. Collected state is merged with the state from the previous steps of the refresh process and with the tracker state at the end.

GitOrigin-RevId: 7aaee23e4030ca151251c04ecbb4557ce3b2ce9f