[workspace model] fix deadlock in LegacyBridgeModulesTest
If project is loaded on EDT, LegacyBridgeModuleManagerComponent blocks EDT until modules are loaded in separate threads, and this may cause a deadlock if initialization of a module wants to run something on EDT (e.g. to fire events from refreshAndFindFileByPath called from createModuleInstance). So the test is changed to avoid loading project on EDT like it works in production.

GitOrigin-RevId: 3df6ef356d6be31cff5196ceb6b9a67229bfdf97