[InspectionGadgets] IDEA-234434 Ignore annotated objects in Type parameter explicitly extends java.lang.Object inspection
Sometimes It is convenient to declare annotations (NotNull, Nullable,
etc.) for java.lang.Object for a type parameter to extend from. IDEA
used to ignore this fact and always reported that such extending is
meaningless.

This patch adds a checkbox to TypeParameterExtendsObjectInspection which
allows the user to decide if they want to disable the inspection to
report problems for annotated java.lang.Object in type parameter
boundaries.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 72ff25d4a44ef21d16b80cbb0e4c22f5f36fa837