IDEA-237378 [a11] action labels are not accessible in the welcome frame
1. When the DKFM is replaced the original PropertyChangeListener's get lost.
For instance - sun.lwawt.macosx.CAccessibility.CAccessibility - installs a "focusOwner" listener at the early start. So, the listeners should be transferred to the new KFM.

2. AccessibleJActionLinkPanel delegates to myActionLink.getAccessibleContext() where myActionLink is a child of JActionLinkPanel instance. When JActionLinkPanel instance gets focused, the following method is called on its context - getAccessibleChildrenCount(). The method returns -1 as myActionLink has no children. The JActionLinkPanel itself is an "ignored" a11y element so it is simply skipped.

In the fix, a helper AccessibleJComponent inner anonymous class is created, which has (it's important) JActionLinkPanel as its enclosing class, not AccessibleJActionLinkPanel (because AccessibleJComponent is declared as an inner class in a Container). Then myAccessibleHelper.getAccessibleChildrenCount() delegates to the enclosing class instance and asks for its children (myActionLink).

GitOrigin-RevId: 34ba693afaba9bb863581d9fc5f4b750cf06bca1