IDEA-CR-61387: PY-30282 EA-209938 Don't render Python docstrings in external process under read action
The helper script for rendering it is now launched synchronously in a pooled
thread while the origin thread holding the read-write lock is constantly
checking whether it was cancelled by incoming write action. If it was, it
terminates preliminary and later rescheduled again thanks to the capabilities
of NonBlockingReadAction. The pooled thread meanwhile will keep running waiting
for the output of the script (up to 5 seconds timeout). Normally it shouldn't
happen, though, and at least Quick Documentation no longer affects the editor
responsiveness. It's a workaround for the current platform API limitation
that queries ordinary documentation providers under read action with no way to
opt out.

GitOrigin-RevId: bae2ee26e0d9599e63bde90a00aa156a472cdb1b