Avoid setting values for invisible variables in DfaAssist
This could be harmless in most of the cases (though adds an overhead) but if we have a visible variable with the same name we add a condition that both are equal to the same value. And if the types of the variables are incompatible, DFA TypeConstraints solver deduces that the only possible variable value is 'null'.
Fixes IDEA-237222 Strange NPE suggestion from debugger dfa

GitOrigin-RevId: 84bd6f56444321758cf6d0a06936d497d09aa03b