[symbol] api/impl: add `SearchTarget` as a way to abstract `Symbol` from what actually will be searched
- Find/Show Usages actions don't need to know about symbols anymore, they only care about something which has
the presentation and the search implementation, and this something is the `SearchTarget`.

- Several symbols might have the same search target.
Example: Kotlin property `foo` is visible in JVM as a getter+setter, and Groovy has expressions like `foo += 42` which
have a reference to both the getter and the setter. In this case both getter and setter symbols should have equal search
targets, so there will be no disambiguation popup when user will invoke Find/Show Usages on such reference.

GitOrigin-RevId: 2addc0947bf9af64c6f0362ef04d2d6f5ee5e7e9