IDEA-CR-61853: PY-41642 Don't rely on Python version when importing `_unpack_args`, try and catch error instead
The `_unpack_args` function doesn't seem to be available in early versions of Python 3. So it's better not to rely on the Python version check and just try importing the function and execute the fallback code inside the except clause.

(cherry picked from commit 0b5125e3c00602372528ae545e9d5677939619ee)

GitOrigin-RevId: 1f705302f3ae8320cdaf826095b1f0c22ffc73f8