introduce an option to turn ProgressManager#run behavior to allow running tasks in parallel in headless mode.
Running tasks in headless mode normally works differently than in the IDE GUI mode. It is ok for many scenarios, but from the other hand, it may lead to deadlock in totally correct code which is written with the default IDE UX behaviour in mind. Settings this property allows a workaround.

In fact we have many attempts to cast ProgressIndicator to ProgressIndicatorBase on our code. Turning on the parallel tasks execution in headless mode yields class cast exceptions, that is the reason we change the default ProgressIndicator instance from EmptyProgressIndicator to ProgressIndicatorBase.

All changes are activated only if the specific property is set. No changes were introduced for the default behaviour or IDE, headless mode or tests.

GitOrigin-RevId: 6c4adafc4ab8e3750e7c0a4984a134e50b20f7a0