fix cases of RecursionManager both over- and under-memoizing (needed for IDEA-236961)
Track preventions per each doPreventingRecursion frame separately to avoid misses when the same prevention happens via different routes.

This increased the number of same memoizations with different dependencies and made keeping them all in the memory very expensive. But it seems that well-behaved memoized values may only vary depending on where the recursion strongly-connected graph was entered, so we now memoize just one value per key and drop it if that entrance's changes.

GitOrigin-RevId: 8ef0402509af0f300318a2fc85d876b0678474c9