[workspace model] support serialization in JPS format for projects imported from external systems
When a project is imported from Maven/Gradle we store its configuration files in JPS format under the caches directory by default; also there is an option to store them in regular .idea/.iml files. We need to support this in legacy bridge to ensure that JPS build works for such projects.

This change introduces JpsImportedEntitySource which is used to distinguish imported entities, provides an implementation of ExternalSystemModulePropertyManager which stores data in the entity stores, added properties to store module type and other module options in the entity storage. Also it changes expected data for some tests on JPS serialization to make them more consistent with the actual behavior in production.

GitOrigin-RevId: 1e843fc2785b9b0a151ac2070d10be6f0b67edf2