RUBY-16609 consider injected and template languages when creating a dummy newline marker
Fixes CreateASFunctionIntentionTest & PostfixTemplateInUltimateTest.

There was an issue where creating a newline range marker would use the comment syntax of the file's language, and not the injected/template language. This led to invalid syntax in the psi tree (e.g., for javascript injected into XML) and subsequently invalid formatting.

Now when creating the newline marker we use the language of the injected/template element if present, otherwise the file's language.

GitOrigin-RevId: faec526c9efe27754e671faabc96c470150936fd