Indexes: fix indexing deadlock caused by a race condition in LimitedCachedFileContentQueue (IDEA-238381).
Initial implementation of the LimitedCachedFileContentQueue was buggy for the corner case: when the last file is about to be pulled from the queue, a thread A can prematurely get `null` although this file is being loaded by thread B, which later decides to push that file back to the queue, meaning that `null` returned to thread A was incorrect. It made some indexing jobs prematurely end. Bytes reserved for the file in CachedFileLoadLimiter were never released because no thread would come to this queue anymore, assuming the indexing had finished. Leakage of the bytes led to global load limiter overflow, which prevented any new files from being loaded.

GitOrigin-RevId: bce76c37a4b5a59d6ae43d04cb7da2454e1bca6f