[vcs] IDEA-235820: Don't modify TreeModelBuilder using plugins in background
Plugin which modified TreeModelBuilder couldn't be unloaded in some cases. It happened when refreshView was called from background and at the same time plugin was tried to unload. So the background task was running, when unloading task was started (on EDT). Background task invoked invokeLaterIfNeeded and put it to the event queue (after the unloading task), so plugin's data was closed in this lambda and plugin couldn't be unloaded.

GitOrigin-RevId: 4bfd0a07463de506e8d5662f9f9a3a697181a689