IDEA-238805 The classloader reference in a cached disabled icon prevents unloading of plugins
CachedImageIcon has a field myResolver with the concrete type MyUrlResolver.
That class has a field myClassLoader that is set to the plugin class loader.
When the plugin creates a disabled icon for any reason, the plugin cannot be
unloaded anymore.

In oder to fix this, the existing IconLoader.detachClassLoader method is modified
to remove the cached icon as well.